+++ Indicators/QuantConnect.Indicators.csproj
@@ -101,6 +101,7 @@
     <Compile Include="AccelerationBands.cs" />
     <Compile Include="AccumulationDistribution.cs" />
     <Compile Include="AccumulationDistributionOscillator.cs" />
+    <Compile Include="AdvanceDeclineIndicator.cs" />
     <Compile Include="AdvanceDeclineVolumeRatio.cs" />
     <Compile Include="AdvanceDeclineRatio.cs" />
     <Compile Include="ArmsIndex.cs" />
